Rename old krb5.ini files away
We want to always use a new krb5.ini, but it is rude to delete
the old file which may potentially be the only copy a user has of
their local changes.  Rename it to a non-functioning name that
indicates it is no longer being used, so that the old version may
be consulted if needed.
Only attempt the rename if we found one, and ignore errors since
this is not a critical part of the installation.