Kill running processes on upgrades/uninstalls
The InstallValidate action of the windows installer will bring up
a dialog informing us that some currently running processes must
be terminated before installation may proceed, and offers to do so,
but does not actually kill the processes.  We have our own code to
do this which was not running, for two reasons: it was sequenced
after InstallValidate, and we did not have a current list of
processes to look for.
Use our own process-killing code since it actually works.